task 2

+++ python_variable_annotations/2-floor.py
@@ -3,6 +3,6 @@
 """ a type-annotated function floor which takes a float 
 n as argument and returns the floor of the float"""
 
-def floor(n: int) -> float:
+def floor(n: float) -> int:
       """returns n"""
-      return n
+      return int(n)
